London Underground Palette

A palette for the London Underground lines, based on the official palette.

underground_colours

underground_colors

underground_lines

Format

An object of class character of length 13.

Details

The same yellow as for the Circle Line is used for the combination of the Circle and Hammersmith & City lines.

For 'All Lines', the pink colour of the Transport for London Visitor Centre is used. underground_lines is the complement of underground_colours (swapping the names and values), for using as the labels argument to ggplot2::scalel_*_identity().

Examples

underground_colours

#>          All Lines           Bakerloo            Central             Circle 
#>          "#DC006B"          "#B26300"          "#DC241F"          "#FFD329" 
#>       Circle + H&C           District Hammersmith & City            Jubilee 
#>          "#FFD32A"          "#007D32"          "#F4A9BE"          "#A1A5A7" 
#>       Metropolitan           Northern         Piccadilly           Victoria 
#>          "#9B0058"          "#000000"          "#0019A8"          "#0098D8" 
#>    Waterloo & City 
#>          "#93CEBA" 

underground_colors

#>          All Lines           Bakerloo            Central             Circle 
#>          "#DC006B"          "#B26300"          "#DC241F"          "#FFD329" 
#>       Circle + H&C           District Hammersmith & City            Jubilee 
#>          "#FFD32A"          "#007D32"          "#F4A9BE"          "#A1A5A7" 
#>       Metropolitan           Northern         Piccadilly           Victoria 
#>          "#9B0058"          "#000000"          "#0019A8"          "#0098D8" 
#>    Waterloo & City 
#>          "#93CEBA" 

underground_lines

#>              #DC006B              #B26300              #DC241F 
#>          "All Lines"           "Bakerloo"            "Central" 
#>              #FFD329              #FFD32A              #007D32 
#>             "Circle"       "Circle + H&C"           "District" 
#>              #F4A9BE              #A1A5A7              #9B0058 
#> "Hammersmith & City"            "Jubilee"       "Metropolitan" 
#>              #000000              #0019A8              #0098D8 
#>           "Northern"         "Piccadilly"           "Victoria" 
#>              #93CEBA 
#>    "Waterloo & City" 

# NOT RUN {
  underground %>%
    filter(metric == "Train delays longer than 15 minutes",
           year == "2016/17",
           is.na(fourweek),
           is.na(quarter),
           line != "All Lines") %>%
    mutate(fill= underground_colours[line]) %>%
    select(line, value, fill) %>%
    ggplot(aes(line, value, fill = fill)) +
    geom_bar(stat = "identity") +
    scale_fill_identity("", labels = underground_lines, guide = "legend") +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    xlab("") +
    ylab("") +
    ggtitle("Train delays longer than 15 minutes (2016/17)")
# }